vehicle reorg: move movement functions into their own file
move all the movement and collision related functions into their own
file.
